Refactored the tests
-- Moved common fixtures into conftest.py
-- Split the tests approximately according to functionality
-- Expanded the glob, walk, isdir, isfile, and listings tests
