+++ test/resources/f2_xml_response.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="utf-8"?>
+<D:multistatus
+  xmlns:D="DAV:"
+  xmlns:ns1="http://apache.org/dav/props/"
+  xmlns:ns0="DAV:">
+  <D:response
+    xmlns:lp1="DAV:"
+    xmlns:lp2="http://apache.org/dav/props/"
+    xmlns:lp3="LCGDM:">
+    <D:href>/foo/bar/folder2/</D:href>
+    <D:propstat>
+      <D:prop>
+        <lp1:getcontentlength>4096</lp1:getcontentlength>
+        <lp1:getlastmodified>Tue, 25 Feb 2025 16:45:43 GMT</lp1:getlastmodified>
+        <lp1:resourcetype>
+          <D:collection/>
+        </lp1:resourcetype>
+        <lp1:iscollection>1</lp1:iscollection>
+        <lp1:executable>T</lp1:executable>
+        <lp1:iscollection>1</lp1:iscollection>
+      </D:prop>
+      <D:status>HTTP/1.1 200 OK</D:status>
+    </D:propstat>
+  </D:response>
+  <D:response
+    xmlns:lp1="DAV:"
+    xmlns:lp2="http://apache.org/dav/props/"
+    xmlns:lp3="LCGDM:">
+    <D:href>/foo/bar/folder2/file1.md</D:href>
+    <D:propstat>
+      <D:prop>
+        <lp1:getcontentlength>2262</lp1:getcontentlength>
+        <lp1:getlastmodified>Tue, 25 Feb 2025 16:45:43 GMT</lp1:getlastmodified>
+        <lp1:iscollection>0</lp1:iscollection>
+        <lp1:executable>F</lp1:executable>
+      </D:prop>
+      <D:status>HTTP/1.1 200 OK</D:status>
+    </D:propstat>
+  </D:response>
+  <D:response
+    xmlns:lp1="DAV:"
+    xmlns:lp2="http://apache.org/dav/props/"
+    xmlns:lp3="LCGDM:">
+    <D:href>/foo/bar/folder2/file2.md</D:href>
+    <D:propstat>
+      <D:prop>
+        <lp1:getcontentlength>300</lp1:getcontentlength>
+        <lp1:getlastmodified>Tue, 25 Feb 2025 16:45:43 GMT</lp1:getlastmodified>
+        <lp1:iscollection>0</lp1:iscollection>
+        <lp1:executable>F</lp1:executable>
+      </D:prop>
+      <D:status>HTTP/1.1 200 OK</D:status>
+    </D:propstat>
+  </D:response>
+</D:multistatus>
